refactored filter, wishlist, profile, box-sizing, and removed the photoURL from database
