+++ src/index.css
@@ -46,7 +46,7 @@
 
 .game-grid {
   display: grid;
-  grid-template-columns: 2fr 2fr;
+  grid-template-columns: 2fr 2fr 2fr;
   gap: 1rem;
   margin: 0;
   padding: 0;
