+++ src/pages/Wishlist.jsx
@@ -1,8 +1,45 @@
-import ComplexState from "../components/ComplexState";
+import { useState, useEffect } from "react";
 import FetchWishlist from "../components/FetchWishlist";
 
 const Wishlist = () => {
-  return <div>{/*   <ComplexState></ComplexState> */}</div>;
+  const [wishlist, setWishlist] = useState([]);
+  console.log(wishlist);
+
+  var user = JSON.parse(localStorage.getItem("user"));
+
+  useEffect(() => {
+    fetch(
+      `${import.meta.env.VITE_API_URL}/users/wishlist/` + user.id,
+
+      {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        credentials: "include",
+      }
+    )
+      .then((response) => response.json())
+      .then((data) => setWishlist(data));
+  }, []);
+
+  return (
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+        flexDirection: "column",
+      }}
+    >
+      <h1>Önskelista</h1>
+      {wishlist.map((g, i) => {
+        return (
+          <FetchWishlist key={i} title={g.title} price={g.price} id={g.id} />
+        );
+      })}
+    </div>
+  );
 };
 
 export default Wishlist;
